requirements.txt
python-dotenv==1.0.1
langchain==0.1.11
langchain-community==0.0.25
